Prepare support for two PC Speaker models
