+++ src/hardware/pcspeaker.cpp
@@ -1,9 +1,7 @@
 /*
  *  SPDX-License-Identifier: GPL-2.0-or-later
  *
- *  Copyright (C) 2020-2022  The DOSBox Staging Team
- *  Copyright (C) 2011-2011  ripa, from vogons.org
- *  Copyright (C) 2002-2019  The DOSBox Team
+ *  Copyright (C) 2022-2022  The DOSBox Staging Team
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -23,623 +21,38 @@
 // NOTE: a lot of this code assumes that the callback is called every emulated
 // millisecond
 
-#include "dosbox.h"
-
-// #define SPKR_DEBUGGING
-// #define REFERENCE
-#include <algorithm>
-#include <array>
-#include <cassert>
-#include <cfloat>
-#include <cmath>
-#include <deque>
-
-#include "inout.h"
-#include "mixer.h"
-#include "timer.h"
-#include "setup.h"
-#include "soft_limiter.h"
-#include "support.h"
-#include "pic.h"
-#include "checks.h"
-
-CHECK_NARROWING();
-
-class PcSpeaker {
-public:
-	PcSpeaker();
-	~PcSpeaker();
-
-	void SetFilterState(const FilterState &filter_state);
-	void SetCounter(const int cntr, const PitMode pit_mode);
-	void SetPITControl(const PitMode pit_mode);
-	void SetType(const PpiPortB &port_b_state);
-
-private:
-	void AddImpulse(float index, const int16_t amplitude);
-	void AddPITOutput(const float index);
-	void ChannelCallback(uint16_t requested_frames);
-	void ForwardPIT(const float new_index);
-	float CalcImpulse(const double t);
-	void InitializeImpulseLUT();
-
-	// Constants
-	static constexpr char device_name[] = "PCSPEAKER";
-
-	static constexpr int16_t positive_amplitude = 20000;
-	static constexpr int16_t negative_amplitude = -positive_amplitude;
-	static constexpr int16_t neutral_amplitude  = 0;
-
-	static constexpr float ms_per_pit_tick = 1000.0f / PIT_TICK_RATE;
-
-	// Mixer channel constants
-	static constexpr auto sample_rate         = 32000;
-	static constexpr auto lowpass_cutoff_freq = 4300;
-	static constexpr auto sample_rate_per_ms  = sample_rate / 1000;
-	static constexpr auto minimum_counter = 2 * PIT_TICK_RATE / sample_rate;
-
-	// must be greater than 0.0f
-	static constexpr float cutoff_margin = 0.2f;
-
-	// Should be selected based on sampling rate
-	static constexpr float high_pass_amount   = 0.999f;
-	static constexpr auto filter_quality      = 100;
-	static constexpr auto oversampling_factor = 32;
-	static constexpr auto filter_width        = filter_quality * oversampling_factor;
-
-	// Compound types and containers
-	struct PitState {
-		// PIT starts in mode 3 (SquareWave) at ~903 Hz (pit_max) with
-		// positive amplitude
-		float max_ms            = ms_per_pit_tick * 1320;
-		float new_max_ms        = max_ms;
-		float half_max_ms       = max_ms / 2;
-		float new_half_max_ms   = half_max_ms;
-		float index             = 0.0f;
-		float last_index        = index;
-		float mode1_pending_max = 0.0f;
-
-		// PIT boolean state
-		bool mode1_waiting_for_counter = false;
-		bool mode1_waiting_for_trigger = true;
-		bool mode3_counting            = false;
-
-		PitMode mode = PitMode::SquareWave;
-
-		int16_t amplitude      = positive_amplitude;
-		int16_t prev_amplitude = negative_amplitude;
-	} pit = {};
-
-	std::deque<float> waveform_deque = {};
-
-	std::array<float, filter_width> impulse_lut = {};
-
-	std::unique_ptr<SoftLimiter> soft_limiter = {};
-
-	mixer_channel_t channel = {};
-
-	PpiPortB prev_port_b_state = {};
-
-	int tally_of_silence = 0;
-};
-
-void PcSpeaker::AddPITOutput(const float index)
-{
-	if (prev_port_b_state.speakerOutput) {
-		AddImpulse(index, pit.amplitude);
-	}
-}
-
-void PcSpeaker::ForwardPIT(const float new_index)
-{
-	auto passed = new_index - pit.last_index;
-
-	auto delay_base = pit.last_index;
-
-	pit.last_index = new_index;
-
-	switch (pit.mode) {
-	case PitMode::Inactive: return;
-
-	case PitMode::InterruptOnTC:
-		if (pit.index >= pit.max_ms) {
-			/// counter reached zero before previous call so do nothing
-			return;
-		}
-		pit.index += passed;
-		if (pit.index >= pit.max_ms) {
-			// counter reached zero between previous and this call
-			const auto delay = delay_base + pit.max_ms - pit.index + passed;
-			pit.amplitude    = positive_amplitude;
-			AddPITOutput(delay);
-		}
-		return;
-
-	case PitMode::OneShot:
-		if (pit.mode1_waiting_for_counter) {
-			// assert output amplitude is high
-			return; // counter not written yet
-		}
-		if (pit.mode1_waiting_for_trigger) {
-			// assert output amplitude is high
-			return; // no pulse yet
-		}
-		if (pit.index >= pit.max_ms) {
-			// counter reached zero before previous call so do nothing
-			return;
-		}
-		pit.index += passed;
-		if (pit.index >= pit.max_ms) {
-			// counter reached zero between previous and this call
-			const auto delay = delay_base + pit.max_ms - pit.index + passed;
-			pit.amplitude    = positive_amplitude;
-			AddPITOutput(delay);
-			// finished with this pulse
-			pit.mode1_waiting_for_trigger = 1;
-		}
-		return;
-
-	case PitMode::RateGenerator:
-	case PitMode::RateGeneratorAlias:
-		while (passed > 0) {
-			// passed the initial low cycle?
-			if (pit.index >= pit.half_max_ms) {
-				// Start a new low cycle
-				if ((pit.index + passed) >= pit.max_ms) {
-					const auto delay = pit.max_ms - pit.index;
-					delay_base += delay;
-					passed -= delay;
-					pit.amplitude = negative_amplitude;
-					AddPITOutput(delay_base);
-					pit.index = 0;
-				} else {
-					pit.index += passed;
-					return;
-				}
-			} else {
-				if ((pit.index + passed) >= pit.half_max_ms) {
-					const auto delay = pit.half_max_ms - pit.index;
-					delay_base += delay;
-					passed -= delay;
-					pit.amplitude = positive_amplitude;
-					AddPITOutput(delay_base);
-					pit.index = pit.half_max_ms;
-				} else {
-					pit.index += passed;
-					return;
-				}
-			}
-		}
-		break;
-
-	case PitMode::SquareWave:
-	case PitMode::SquareWaveAlias:
-		if (!pit.mode3_counting)
-			break;
-		while (passed > 0) {
-			// Determine where in the wave we're located
-			if (pit.index >= pit.half_max_ms) {
-				if ((pit.index + passed) >= pit.max_ms) {
-					const auto delay = pit.max_ms - pit.index;
-					delay_base += delay;
-					passed -= delay;
-					pit.amplitude = positive_amplitude;
-					AddPITOutput(delay_base);
-					pit.index = 0;
-					// Load the new count
-					pit.max_ms      = pit.new_max_ms;
-					pit.half_max_ms = pit.new_half_max_ms;
-				} else {
-					pit.index += passed;
-					return;
-				}
-			} else {
-				if ((pit.index + passed) >= pit.half_max_ms) {
-					const auto delay = pit.half_max_ms - pit.index;
-					delay_base += delay;
-					passed -= delay;
-					pit.amplitude = negative_amplitude;
-					AddPITOutput(delay_base);
-					pit.index = pit.half_max_ms;
-					// Load the new count
-					pit.max_ms      = pit.new_max_ms;
-					pit.half_max_ms = pit.new_half_max_ms;
-				} else {
-					pit.index += passed;
-					return;
-				}
-			}
-		}
-		break;
-
-	case PitMode::SoftwareStrobe:
-		if (pit.index < pit.max_ms) {
-			// Check if we're gonna pass the end this block
-			if (pit.index + passed >= pit.max_ms) {
-				const auto delay = pit.max_ms - pit.index;
-				delay_base += delay;
-				passed -= delay;
-				pit.amplitude = negative_amplitude;
-				// No new events unless reprogrammed
-				AddPITOutput(delay_base);
-				pit.index = pit.max_ms;
-			} else
-				pit.index += passed;
-		}
-		break;
-	default:
-		// other modes not implemented
-		break;
-	}
-}
-
-void PcSpeaker::SetPITControl(const PitMode pit_mode)
-{
-	const auto new_index = static_cast<float>(PIC_TickIndex());
-	ForwardPIT(new_index);
-#ifdef SPKR_DEBUGGING
-	LOG_INFO("PCSPEAKER: %f pit command: %s",
-	         PIC_FullIndex(),
-	         PitModeToString(pit_mode));
-#endif
-	// TODO: implement all modes
-	switch (pit_mode) {
-	case PitMode::OneShot:
-		pit.mode      = pit_mode;
-		pit.amplitude = positive_amplitude;
-
-		pit.mode1_waiting_for_counter = 1;
-		pit.mode1_waiting_for_trigger = 0;
-		break;
-
-	case PitMode::SquareWave:
-	case PitMode::SquareWaveAlias:
-		pit.mode      = pit_mode;
-		pit.amplitude = positive_amplitude;
-
-		pit.mode3_counting = false;
-		break;
-
-	default: return;
-	}
-	AddPITOutput(new_index);
-}
-
-void PcSpeaker::SetCounter(const int cntr, const PitMode pit_mode)
-{
-#ifdef SPKR_DEBUGGING
-	LOG_INFO("PCSPEAKER: %f counter: %u, mode: %s",
-	         PIC_FullIndex(),
-	         cntr,
-	         PitModeToString(pit_mode));
-#endif
-	const auto new_index = static_cast<float>(PIC_TickIndex());
-
-	const auto duration_of_count_ms = ms_per_pit_tick * static_cast<float>(cntr);
-	ForwardPIT(new_index);
-
-	switch (pit_mode) {
-	case PitMode::InterruptOnTC:
-		// used with "realsound" (PWM)
-		pit.index     = 0;
-		pit.amplitude = negative_amplitude;
-		pit.max_ms    = duration_of_count_ms;
-		AddPITOutput(new_index);
-		break;
-
-	case PitMode::OneShot: // used by Star Control 1
-		pit.mode1_pending_max = duration_of_count_ms;
-		if (pit.mode1_waiting_for_counter) {
-			// assert output amplitude is high
-			pit.mode1_waiting_for_counter = 0;
-			pit.mode1_waiting_for_trigger = 1;
-		}
-		break;
-
-	// Single cycle low, rest low high generator
-	case PitMode::RateGenerator:
-	case PitMode::RateGeneratorAlias:
-		pit.index     = 0;
-		pit.amplitude = negative_amplitude;
-		AddPITOutput(new_index);
-		pit.max_ms      = duration_of_count_ms;
-		pit.half_max_ms = ms_per_pit_tick;
-		break;
-
-	case PitMode::SquareWave:
-	case PitMode::SquareWaveAlias:
-		if (cntr < minimum_counter) {
-			// avoid breaking digger music
-			pit.amplitude = positive_amplitude;
-			pit.mode      = PitMode::Inactive;
-			AddPITOutput(new_index);
-			return;
-		}
-		pit.new_max_ms      = duration_of_count_ms;
-		pit.new_half_max_ms = pit.new_max_ms / 2;
-		if (!pit.mode3_counting) {
-			pit.index       = 0;
-			pit.max_ms      = pit.new_max_ms;
-			pit.half_max_ms = pit.new_half_max_ms;
-			if (prev_port_b_state.timerGateOutput) {
-				pit.mode3_counting = true;
-				// probably not necessary
-				pit.amplitude = positive_amplitude;
-				AddPITOutput(new_index);
-			}
-		}
-		break;
-
-	case PitMode::SoftwareStrobe:
-		pit.amplitude = positive_amplitude;
-		AddPITOutput(new_index);
-		pit.index  = 0;
-		pit.max_ms = duration_of_count_ms;
-		break;
-
-	default:
-#ifdef SPKR_DEBUGGING
-		LOG_MSG("Unhandled speaker mode %s at %f",
-		        PitModeToString(pit_mode),
-		        PIC_FullIndex());
-#endif
-		return;
-	}
-	pit.mode = pit_mode;
-}
-
-void PcSpeaker::SetType(const PpiPortB &port_b_state)
-{
-#ifdef SPKR_DEBUGGING
-	LOG_INFO("PCSPEAKER: %f output: %s, clock gate %s",
-	         PIC_FullIndex(),
-	         port_b_state.speakerOutput ? "pit" : "forced low",
-	         port_b_state.timerGateOutput ? "on" : "off");
-#endif
-	const auto new_index = static_cast<float>(PIC_TickIndex());
-	ForwardPIT(new_index);
-	// pit clock gate enable rising edge is a trigger
-	const bool pit_trigger = !prev_port_b_state.timerGateOutput && port_b_state.timerGateOutput;
-	prev_port_b_state.data = port_b_state.data;
-	if (pit_trigger) {
-		switch (pit.mode) {
-		case PitMode::OneShot:
-			if (pit.mode1_waiting_for_counter) {
-				// assert output amplitude is high
-				break;
-			}
-			pit.amplitude = negative_amplitude;
-			pit.index     = 0;
-			pit.max_ms    = pit.mode1_pending_max;
-
-			pit.mode1_waiting_for_trigger = 0;
-			break;
-
-		case PitMode::SquareWave:
-		case PitMode::SquareWaveAlias:
-			pit.mode3_counting = true;
-			// pit.new_max_ms = pit.new_max_ms; // typo or bug?
-			pit.index           = 0;
-			pit.max_ms          = pit.new_max_ms;
-			pit.new_half_max_ms = pit.new_max_ms / 2;
-			pit.half_max_ms     = pit.new_half_max_ms;
-			pit.amplitude       = positive_amplitude;
-			break;
-
-		default:
-			// TODO: implement other modes
-			break;
-		}
-	} else if (!port_b_state.timerGateOutput) {
-		switch (pit.mode) {
-		case PitMode::OneShot:
-			// gate amplitude does not affect mode1
-			break;
-
-		case PitMode::SquareWave:
-		case PitMode::SquareWaveAlias:
-			// low gate forces pit output high
-			pit.amplitude      = positive_amplitude;
-			pit.mode3_counting = false;
-			break;
-
-		default:
-			// TODO: implement other modes
-			break;
-		}
-	}
-	if (port_b_state.speakerOutput) {
-		AddImpulse(new_index, pit.amplitude);
-	} else {
-		AddImpulse(new_index, negative_amplitude);
-	}
-}
-
-// TODO: check if this is accurate
-static double sinc(const double t)
-{
-	double result = 1.0;
-
-	constexpr auto SINC_ACCURACY = 20;
-	for (auto k = 1; k < SINC_ACCURACY; ++k) {
-		result *= cos(t / pow(2.0, k));
-	}
-	return result;
-}
-
-float PcSpeaker::CalcImpulse(const double t)
-{
-	// raised-cosine-windowed sinc function
-	const double fs = sample_rate;
-	const auto fc   = fs / (2 + static_cast<double>(cutoff_margin));
-	const auto q    = static_cast<double>(filter_quality);
-	if ((0 < t) && (t * fs < q)) {
-		const auto window    = 1.0 + cos(2 * fs * M_PI * (q / (2 * fs) - t) / q);
-		const auto amplitude = window * (sinc(2 * fc * M_PI * (t - q / (2 * fs)))) / 2.0;
-		return static_cast<float>(amplitude);
-	} else
-		return 0.0f;
-}
-
-void PcSpeaker::AddImpulse(float index, int16_t amplitude)
-{
-	// Did the amplitude change?
-	if (amplitude == pit.prev_amplitude)
-		return;
-
-	pit.prev_amplitude = amplitude;
-
-	// Wake the channel if it was sleeping
-	if (!channel->is_enabled)
-		channel->Enable(true);
-
-	// Make sure the time index is valid
-	index = clamp(index, 0.0f, 1.0f);
-#ifndef REFERENCE
-	const auto samples_in_impulse = index * sample_rate_per_ms;
-	auto phase = static_cast<int>(samples_in_impulse * oversampling_factor) %
-	             oversampling_factor;
-
-	auto offset = static_cast<int>(samples_in_impulse);
-	if (phase != 0) {
-		offset++;
-		phase = oversampling_factor - phase;
-	}
-
-	for (int i = 0; i < filter_quality; ++i) {
-		const auto wave_i    = static_cast<size_t>(offset + i);
-		const auto impulse_i = static_cast<size_t>(phase + i * oversampling_factor);
-		waveform_deque.at(wave_i) += amplitude * impulse_lut.at(impulse_i);
-	}
-}
-#else
-	const auto portion_of_ms = static_cast < double(index) / 1000.0;
-	for (size_t i = 0; i < waveform_deque.size(); ++i) {
-		const auto impulse_time = static_cast<double>(i) / sample_rate - portion_of_ms;
-
-		waveform_deque[i] += amplitude * CalcImpulse(impulse_time);
-	}
-}
-#endif
-
-void PcSpeaker::ChannelCallback(uint16_t requested_frames)
-{
-	ForwardPIT(1.0f);
-	pit.last_index = 0;
-
-	constexpr auto num_channels = 2;
-	// Static vectors and samples re-used for all callbacks
-	static std::vector<float> in_buffer(num_channels);
-	static std::vector<int16_t> out_buffer(num_channels);
-	static auto &out_sample = out_buffer.front();
-	static auto &in_sample  = in_buffer.front();
-
-	while (requested_frames > 0 && waveform_deque.size()) {
-		// Pop the first sample off the waveform
-		in_sample += waveform_deque.front();
-		waveform_deque.pop_front();
-		waveform_deque.push_back(0.0f);
-
-		// Soft-limit the sample, if needed
-		soft_limiter->Process(in_buffer, 1, out_buffer);
-
-		// Pass the sample to the mixer and decrement the requested frames
-		channel->AddSamples_m16(1, &out_sample);
-		--requested_frames;
-
-		// Keep a tally of sequential silence so we can sleep the channel
-		tally_of_silence = out_sample ? 0 : tally_of_silence + 1;
-
-		// Scale down the running volume amplitude. Eventually it will
-		// hit 0 if no other waveforms are generated.
-		in_sample *= high_pass_amount;
-	}
-
-	// Write silence if the waveform deque ran out
-	while (requested_frames) {
-		channel->AddSamples_m16(1, &neutral_amplitude);
-		++tally_of_silence;
-		--requested_frames;
-	}
-
-	// Maybe put the channel to sleep after 10s
-	if (tally_of_silence > 100000) {
-		pit.prev_amplitude = neutral_amplitude;
-		channel->Enable(false);
-		tally_of_silence = 0;
-	}
-}
-
-void PcSpeaker::InitializeImpulseLUT()
-{
-	assert(impulse_lut.size() == filter_width);
-	for (auto i = 0u; i < filter_width; ++i)
-		impulse_lut[i] = CalcImpulse(i / (static_cast<double>(sample_rate) * oversampling_factor));
-}
-
-void PcSpeaker::SetFilterState(const FilterState &filter_state)
-{
-	assert(channel);
-	
-	// Setup the low-pass filter
-	if (filter_state == FilterState::On) {
-		// The filters are meant to emulate the bandwidth limited
-		// sound of the small PC speaker. This more accurately
-		// reflects people's actual experience of the PC speaker
-		// sound than the raw unfiltered output, and it's a lot
-		// more pleasant to listen to, especially in headphones.
-		constexpr auto hp_order       = 3;
-		constexpr auto hp_cutoff_freq = 120;
-		channel->ConfigureHighPassFilter(hp_order, hp_cutoff_freq);
-		channel->SetHighPassFilter(FilterState::On);
-
-		constexpr auto lp_order       = 2;
-		constexpr auto lp_cutoff_freq = 4800;
-		channel->ConfigureLowPassFilter(lp_order, lp_cutoff_freq);
-		channel->SetLowPassFilter(FilterState::On);
-	} else {
-		channel->SetHighPassFilter(FilterState::Off);
-		channel->SetLowPassFilter(FilterState::Off);
-	}
-}
-
-PcSpeaker::PcSpeaker()
-{
-	static_assert(sample_rate >= 8000, "Sample rate must be at least 8 kHz");
-	static_assert(sample_rate % 1000 == 0, "PC Speaker sample must be a multiple of 1000");
-
-	InitializeImpulseLUT();
-
-	// Size the waveform queue
-	constexpr auto waveform_size = filter_quality + sample_rate_per_ms;
-	waveform_deque.resize(check_cast<uint16_t>(waveform_size), 0.0f);
-
-	// Setup the soft limiter
-	soft_limiter = std::make_unique<SoftLimiter>(device_name);
-	assert(soft_limiter);
-
-	// Register the sound channel
-	const auto callback = std::bind(&PcSpeaker::ChannelCallback, this, std::placeholders::_1);
-
-	channel = MIXER_AddChannel(callback,
-	                           sample_rate,
-	                           device_name,
-	                           {ChannelFeature::ReverbSend, ChannelFeature::ChorusSend});
-	assert(channel);
-
-	LOG_MSG("%s: Initialized", device_name);
-
-	channel->SetPeakAmplitude(static_cast<uint32_t>(positive_amplitude));
-}
-
-PcSpeaker::~PcSpeaker()
-{
-	channel->Enable(false);
-	LOG_MSG("%s: Shutting down", device_name);
-}
+#include "pcspeaker_impulse.h"
 
 // The PC Speaker managed pointer
 std::unique_ptr<PcSpeaker> pc_speaker = {};
+
+void PCSPEAKER_ShutDown([[maybe_unused]] Section *sec)
+{
+	pc_speaker.reset();
+}
+
+void PCSPEAKER_Init(Section *section)
+{
+	assert(section);
+	const auto prop = static_cast<Section_prop *>(section);
+
+	if (!prop->Get_bool("pcspeaker"))
+		return;
+
+	// Create the PC Speaker
+	if (!pc_speaker)
+		pc_speaker = std::make_unique<PcSpeakerImpulse>();
+
+	// Get the user's filering preference
+	const auto filter_pref  = std::string_view(prop->Get_string("pcspeaker_filter"));
+	const auto filter_state = filter_pref == "on" ? FilterState::On : FilterState::Off;
+
+	// Apply the user's filtering preference
+	assert(pc_speaker);
+	pc_speaker->SetFilterState(filter_state);
+
+	section->AddDestroyFunction(&PCSPEAKER_ShutDown, true);
+}
 
 // PC Speaker external API, used by the PIT timer and keyboard
 void PCSPEAKER_SetCounter(const int counter, const PitMode pit_mode)
@@ -658,32 +71,4 @@ void PCSPEAKER_SetType(const PpiPortB &port_b_state)
 {
 	if (pc_speaker)
 		pc_speaker->SetType(port_b_state);
-}
-
-void PCSPEAKER_ShutDown([[maybe_unused]] Section *sec)
-{
-	pc_speaker.reset();
-}
-
-void PCSPEAKER_Init(Section *section)
-{
-	assert(section);
-	const auto prop = static_cast<Section_prop *>(section);
-
-	if (!prop->Get_bool("pcspeaker"))
-		return;
-
-	// Create the PC Speaker
-	if (!pc_speaker)
-		pc_speaker = std::make_unique<PcSpeaker>();
-
-	// Get the user's filering preference
-	const auto filter_pref  = std::string_view(prop->Get_string("pcspeaker_filter"));
-	const auto filter_state = filter_pref == "on" ? FilterState::On : FilterState::Off;
-
-	// Apply the user's filtering preference
-	assert(pc_speaker);
-	pc_speaker->SetFilterState(filter_state);
-
-	section->AddDestroyFunction(&PCSPEAKER_ShutDown, true);
 }
