+++ vs/dosbox.vcxproj.filters
@@ -199,6 +199,9 @@
     <ClCompile Include="..\src\hardware\pcspeaker.cpp">
       <Filter>src\hardware</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\hardware\pcspeaker_impulse.cpp">
+      <Filter>src\hardware</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\hardware\pic.cpp">
       <Filter>src\hardware</Filter>
     </ClCompile>
